requirements.txt
Flask
langchain-google-genai
langchain-community
faiss-cpu
pdf2image
python-dotenv
Pillow
